Finish

+++ Lab18/Form1.cs
@@ -1,12 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
 using System.IO;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using System.Windows.Forms;
 
 namespace Lab18
@@ -14,14 +9,14 @@ namespace Lab18
     public partial class Form1 : Form
     {
         private const string FolderName = "Folder_";
-        
+
         private const string TestDirectoryName = "100 Folders here!";
-        
+
         private const int NumberSubFolders = 100;
-        
-        private int _subFoldersCount;
-        
+
         private readonly DirectoryInfo _dirInfo;
+
+        private int _subFoldersCount;
 
         public Form1()
         {
@@ -35,6 +30,7 @@ namespace Lab18
             foreach (var folder in _dirInfo.GetDirectories())
                 folder.Delete(true);
         }
+
         private void ShowMaxNumberSubFolder()
         {
             var path = "D:";
@@ -54,9 +50,9 @@ namespace Lab18
                 MessageBox.Show("Максимальное количество папок " + count);
             }
 
-            Directory.Delete(@"D:\a",true);
+            Directory.Delete(@"D:\a", true);
         }
-        
+
         private void CreateSubFolders(DirectoryInfo dir)
         {
             try
@@ -75,25 +71,17 @@ namespace Lab18
 
         private void CreateFolder_Click(object sender, EventArgs e)
         {
-            if (!_dirInfo.Exists)
-            {
-                _dirInfo.Create();
-            }
+            if (!_dirInfo.Exists) _dirInfo.Create();
 
-            for (int i = 1; i <= 100; i++)
-            {
-                _dirInfo.CreateSubdirectory(FolderName + i);
-            }
+            for (var i = 1; i <= 100; i++) _dirInfo.CreateSubdirectory(FolderName + i);
 
             MessageBox.Show("Было создано 100 папок на рабочем столе. ");
 
-            if (_dirInfo.Exists)
-            {
-                DeleteFolders();
-            }
+            if (_dirInfo.Exists) DeleteFolders();
 
             _dirInfo.Delete(true);
         }
+
         private void CreateSubFolder_Click(object sender, EventArgs e)
         {
             if (!_dirInfo.Exists)
@@ -103,6 +91,7 @@ namespace Lab18
 
             CreateSubFolders(_dirInfo);
         }
+
         private void CheckMaxSubFolders_Click(object sender, EventArgs e)
         {
             ShowMaxNumberSubFolder();
@@ -111,15 +100,12 @@ namespace Lab18
         private void StartFindButton_Click(object sender, EventArgs e)
         {
             if (!string.IsNullOrEmpty(TextBoxFind.Text) && !string.IsNullOrEmpty(TextBoxFindPlace.Text))
-            {
-                foreach (string path in EnumerateAllFiles(TextBoxFindPlace.Text, TextBoxFind.Text))
-                {
+                foreach (var path in EnumerateAllFiles(TextBoxFindPlace.Text, TextBoxFind.Text))
                     if (MessageBox.Show($"Показать содержимое этого файла: {path}", "Просмотр файла:",
                         MessageBoxButtons.YesNo) == DialogResult.Yes)
-                    {
                         try
                         {
-                            using (StreamReader streamReader = new StreamReader(path))
+                            using (var streamReader = new StreamReader(path))
                             {
                                 TextFile.Text = streamReader.ReadToEnd();
                             }
@@ -129,20 +115,17 @@ namespace Lab18
                             MessageBox.Show("Error ;(");
                             throw;
                         }
-                    }
-                }
-            }
         }
 
         private void TextBoxFindPlace_Click(object sender, EventArgs e)
         {
-            FolderBrowserDialog browserDialog = new FolderBrowserDialog();
+            var browserDialog = new FolderBrowserDialog();
             if (browserDialog.ShowDialog() == DialogResult.OK)
                 TextBoxFindPlace.Text = browserDialog.SelectedPath;
         }
 
         private static IEnumerable<string> EnumerateAllFiles(string path, string pattern)
-        { 
+        {
             IEnumerable<string> files = null;
             try
             {
@@ -154,9 +137,8 @@ namespace Lab18
             }
 
             if (files != null)
-            {
-                foreach (var file in files) yield return file;
-            }
+                foreach (var file in files)
+                    yield return file;
 
             IEnumerable<string> directories = null;
             try
@@ -170,10 +152,7 @@ namespace Lab18
 
             if (directories == null) yield break;
             {
-                foreach (var file in directories.SelectMany(d => EnumerateAllFiles(d, pattern)))
-                {
-                    yield return file;
-                }
+                foreach (var file in directories.SelectMany(d => EnumerateAllFiles(d, pattern))) yield return file;
             }
         }
 
@@ -182,7 +161,7 @@ namespace Lab18
             if (string.IsNullOrEmpty(FileNameTextBox.Text)) return;
             try
             {
-                using (StreamReader streamReader = new StreamReader(FileNameTextBox.Text))
+                using (var streamReader = new StreamReader(FileNameTextBox.Text))
                 {
                     TextFile.Text = streamReader.ReadToEnd();
                 }
@@ -196,25 +175,19 @@ namespace Lab18
 
         private void FileNameTextBox_Click(object sender, EventArgs e)
         {
-            OpenFileDialog fileDialog = new OpenFileDialog();
-            if (fileDialog.ShowDialog() == DialogResult.OK)
-            {
-                FileNameTextBox.Text = fileDialog.FileName;
-            }
+            var fileDialog = new OpenFileDialog();
+            if (fileDialog.ShowDialog() == DialogResult.OK) FileNameTextBox.Text = fileDialog.FileName;
         }
 
         private int TryCreateFolder()
         {
             var tempName = "a";
-            
+
             try
             {
                 while (true)
                 {
-                    if (!_dirInfo.Exists)
-                    {
-                        _dirInfo.Create();
-                    }
+                    if (!_dirInfo.Exists) _dirInfo.Create();
                     _dirInfo.CreateSubdirectory(tempName);
                     tempName += "a";
                     _dirInfo.Delete(true);
@@ -224,10 +197,11 @@ namespace Lab18
             {
                 return tempName.Length - 1;
             }
-            
         }
 
-        private void CheckFolderNameLengthButton_Click(object sender, EventArgs e) => MessageBox.Show(TryCreateFolder().ToString());
-        
+        private void CheckFolderNameLengthButton_Click(object sender, EventArgs e)
+        {
+            MessageBox.Show(TryCreateFolder().ToString());
+        }
     }
 }
